Beta v0.3: - Исправлен экспорт PDF/Excel - Добавлена возможность редактирования замен - Кнопки взаимодействия с заменой вынесены в подменю - Система согласования замен. - Связь дисциплин и групп, соответствующие проверки.

+++ kmpo_web/functions/delete_replacement.php
@@ -8,20 +8,30 @@ header('Content-Type: application/json');
 // Получение ID замены
 $replacement_id = $_GET['replacement_id'];
 
-if ($replacement_id){
-    try {
-        $url = "http://localhost:305/replace_delete";
-        $data = ['replacement_id' => $replacement_id];
-        $response = sendPostRequest($url, $data);
 
+if ($replacement_id){
+    if($DB->isConfirmed($replacement_id) == 1){
+        try {
+            $url = "http://localhost:305/replace_delete";
+            $data = ['replacement_id' => $replacement_id];
+            $response = sendPostRequest($url, $data);
+    
+            if (!$DB->deleteReplacement($replacement_id)) {
+                echo json_encode(['success' => false, 'message' => 'Возникла ошибка при удалении замены из БД!']);
+            }else{
+                echo json_encode(['success' => true, 'message' => "Замена №$replacement_id была удалена!"]);
+            }
+        } catch (Exception $e) {
+            echo json_encode(['success' => false, 'message' => $e->getMessage()]);
+        }  
+    }
+    else{
         if (!$DB->deleteReplacement($replacement_id)) {
             echo json_encode(['success' => false, 'message' => 'Возникла ошибка при удалении замены из БД!']);
         }else{
             echo json_encode(['success' => true, 'message' => "Замена №$replacement_id была удалена!"]);
         }
-    } catch (Exception $e) {
-        echo json_encode(['success' => false, 'message' => $e->getMessage()]);
-    }    
+    }
 }
 else{
     echo json_encode(['success' => false, 'message' => 'Не указан ID замены!']);
